refactor(brski): rename C++ headers to `.hpp`
Rename C++ headers to `.hpp` instead of `.h`, to avoid confusion with
C headers, which are also named `.h`.
