+++ src/brski/http/httplib_wrapper.hpp
@@ -12,7 +12,7 @@
 #define HTTPLIB_WRAPPER_H
 
 #include <vector>
-#include "http.h"
+#include "http.hpp"
 
 /**
  * @brief Starts the http library server
