+++ src/brski/http/https_server.hpp
@@ -12,7 +12,7 @@
 #define HTTPS_SERVER_H
 #include <vector>
 
-#include "http.h"
+#include "http.hpp"
 
 /**
  * @brief Starts the https server
